+++ packages/ui-core/.eslintrc.cjs
@@ -1,0 +1,10 @@
+module.exports = {
+  extends: '../../.eslintrc.cjs',
+  settings: {
+    'import/resolver': {
+      typescript: {
+        project: './tsconfig.json',
+      },
+    },
+  },
+}
