+++ packages/ui-core/README.md
@@ -1,0 +1,3 @@
+# `@hedron/ui-core`
+
+A collection of components and hooks used by Hedron desktop as well as any related plugins/inputs.
