+++ packages/ui-core/src/components/Button/Button.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
+import { Icon, IconName } from '@components/Icon/Icon'
 import c from './Button.module.css'
-import { Icon, IconName } from '@components/core/Icon/Icon'
 
 export interface ButtonProps extends React.HTMLAttributes<HTMLButtonElement> {
   type?: 'primary' | 'secondary' | 'neutral' | 'ghost' | 'danger'
