+++ packages/ui-core/src/components/Dialog/Dialog.module.css
