+++ packages/ui-core/src/components/FloatSlider/FloatSlider.module.css
