+++ packages/ui-core/src/components/FloatSlider/FloatSlider.tsx
