+++ packages/ui-core/src/components/Panel/Panel.module.css
