+++ packages/ui-core/src/components/SideTabs/SideTabs.module.css
