+++ packages/ui-core/src/components/SideTabs/SideTabs.tsx
