+++ packages/ui-core/src/stories/Button.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { Button } from '@components/core/Button/Button'
+import { Button } from '@components/Button/Button'
 
 const meta = {
   title: 'Button',
