+++ packages/ui-core/src/stories/ClockDisplay.stories.ts
